requirements.txt
# Core dependencies
httpx>=0.25.0
click>=8.1.0
pyyaml>=6.0
tabulate>=0.9.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0